+++ pkg/cmd/task/logs.go
@@ -24,7 +24,7 @@ import (
 	"github.com/tektoncd/cli/pkg/flags"
 	"github.com/tektoncd/cli/pkg/options"
 	thelper "github.com/tektoncd/cli/pkg/task"
-	trlist "github.com/tektoncd/cli/pkg/helper/taskrun/list"
+	trlist "github.com/tektoncd/cli/pkg/taskrun/list"
 	validate "github.com/tektoncd/cli/pkg/validate"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
