+++ pkg/cmd/taskrun/logs.go
@@ -21,9 +21,9 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/tektoncd/cli/pkg/cli"
-	trlist "github.com/tektoncd/cli/pkg/helper/taskrun/list"
 	"github.com/tektoncd/cli/pkg/log"
 	"github.com/tektoncd/cli/pkg/options"
+	trlist "github.com/tektoncd/cli/pkg/taskrun/list"
 	validate "github.com/tektoncd/cli/pkg/validate"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
