+++ pkg/log/pipeline_reader.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/tektoncd/cli/pkg/helper/pipelinerun"
-	trh "github.com/tektoncd/cli/pkg/helper/taskrun"
+	trh "github.com/tektoncd/cli/pkg/taskrun"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
